requirements.txt
pandas~=2.0.3
numpy~=1.25.1
matplotlib~=3.7.2
missingno~=0.5.2
seaborn~=0.12.2
scipy~=1.11.1
plotly~=5.15.0
squarify~=0.4.3
joblib~=1.3.1
streamlit~=1.25.0
Pillow~=9.5.0
scikit-learn~=1.3.0